requirements_streamlit.txt
streamlit>=1.28.0
pandas>=2.0.0
pydub>=0.25.1
soundfile>=0.12.1
openai-whisper>=20231117
transformers>=4.35.0
groq>=0.4.0
jiwer>=3.0.0
rouge-score>=0.1.2
nltk>=3.8.1
reportlab>=4.0.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
torch>=2.0.0
